Fix para el arco que yo mergeare para que no se pierda
