+++ recreaacademy/src/components/ArcoRecrea/ArcoRecrea.css
@@ -1,83 +1,129 @@
+.cointainer {
+  width: 100%;
+  height: 310px;
+  overflow-y: hidden;
+  margin-bottom: 60px;
+  box-shadow: 0 7px 2px -2px rgb(230, 230, 230);
+}
+
 .arco-container {
-    position: relative;
-    width: 700px;
-    height: 350px;
-    margin: 0 auto;
+  position: relative;
+  width: 700px;
+  height: 350px;
+  margin: 0 auto;
+}
+
+.arco-svg {
+  width: 100%;
+  height: 100%;
+  display: block;
+  position: relative;
+  z-index: 1;
+}
+
+.iconos-rotatorios {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 700px;
+  height: 350px;
+  animation: rotate-circle 25s linear infinite;
+  transform-origin: center bottom;
+  z-index: 2;
+}
+
+.icono {
+  position: absolute;
+  background: white;
+  border-radius: 50%;
+  padding: 10px;
+  font-size: 24px;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+  animation: rotate-reverse 25s linear infinite;
+}
+
+/* Posiciones ajustadas para nuevo arco */
+.icono-0 {
+  top: 270px;
+  left: 100px;
+}
+
+.icono-1 {
+  top: 190px;
+  left: 200px;
+}
+
+.icono-2 {
+  top: 120px;
+  left: 340px;
+}
+
+.icono-3 {
+  top: 190px;
+  right: 200px;
+}
+
+.icono-4 {
+  top: 270px;
+  right: 100px;
+}
+
+.texto-centro-arco {
+  position: absolute;
+  top: -50px;
+  left: 50%;
+  transform: translateX(-50%);
+  text-align: center;
+  z-index: 3;
+}
+
+.texto-centro-arco h1 {
+  font-size: 2.7rem !important;
+  font-weight: bold;
+  margin: 0;
+  color: #e11d48;
+  margin-bottom: 0 !important;
+}
+
+.texto-centro-arco h1 span {
+  font-size: 4rem !important;
+  margin-top: 0 !important;
+  color: #f59e0b;
+}
+
+.anio-container {
+  display: flex;
+  justify-content: end;
+  width: 100%;
+}
+.anio {
+  font-size: 0.9rem;
+  font-weight: bold;
+  background: #e11d48;
+  color: white;
+  border-radius: 10px;
+  padding: 4px 12px;
+  display: inline-block;
+  margin-top: 6px;
+  align-self: right;
+}
+
+@keyframes rotate-circle {
+  from {
+    transform: rotate(0deg);
   }
-  
-  .arco-svg {
-    width: 100%;
-    height: 100%;
-    display: block;
-    position: relative;
-    z-index: 1;
+
+  to {
+    transform: rotate(360deg);
   }
-  
-  .iconos-rotatorios {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 700px;
-    height: 350px;
-    animation: rotate-circle 25s linear infinite;
-    transform-origin: center bottom;
-    z-index: 2;
+}
+
+@keyframes rotate-reverse {
+  from {
+    transform: rotate(0deg);
   }
-  
-  .icono {
-    position: absolute;
-    background: white;
-    border-radius: 50%;
-    padding: 10px;
-    font-size: 24px;
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
-    animation: rotate-reverse 25s linear infinite;
+
+  to {
+    transform: rotate(-360deg);
   }
-  
-  /* Posiciones ajustadas para nuevo arco */
-  .icono-0 { top: 270px; left: 100px; }
-  .icono-1 { top: 190px; left: 200px; }
-  .icono-2 { top: 120px; left: 340px; }
-  .icono-3 { top: 190px; right: 200px; }
-  .icono-4 { top: 270px; right: 100px; }
-  
-  .texto-centro-arco {
-    position: absolute;
-    top: 140px;
-    left: 50%;
-    transform: translateX(-50%);
-    text-align: center;
-    z-index: 3;
-  }
-  
-  .texto-centro-arco h1 {
-    font-size: 2.4rem;
-    font-weight: bold;
-    margin: 0;
-    color: #e11d48;
-  }
-  
-  .texto-centro-arco h1 span {
-    color: #f59e0b;
-  }
-  
-  .anio {
-    font-size: 0.9rem;
-    font-weight: bold;
-    background: #e11d48;
-    color: white;
-    border-radius: 10px;
-    padding: 4px 12px;
-    display: inline-block;
-    margin-top: 6px;
-  }
-  
-  @keyframes rotate-circle {
-    from { transform: rotate(0deg); }
-    to   { transform: rotate(360deg); }
-  }
-  
-  @keyframes rotate-reverse {
-    from { transform: rotate(0deg); }
-    to   { transform: rotate(-360deg); }
-  }
+}
