+++ recreaacademy/src/components/ArcoRecrea/ArcoRecrea.jsx
@@ -1,42 +1,56 @@
-import React from 'react';
+import React, { useState } from 'react';
+import Card from '@mui/material/Card';
+import '../../pages/RutasFormativas/RutasFormativas.css'
 import './ArcoRecrea.css';
 
-const iconos = ['school', 'emoji_events', 'volunteer_activism', 'text_fields', 'rocket_launch'];
+const iconos = [
+  { icon: "local_library", title: "Biblioteca", description: "Accede a contenido educativo digital" },
+  { icon: "school", title: "Escuela", description: "Formación continua para docentes" },
+  { icon: "model_training", title: "Entrenamiento", description: "Capacitación práctica en nuevas tecnologías" },
+  { icon: "labs", title: "Laboratorio", description: "Espacios para experimentar e innovar" },
+  { icon: "digital_wellbeing", title: "Bienestar Digital", description: "Equilibrio y salud en el entorno digital" },
+  { icon: "network_intel_node", title: "Redes", description: "Conecta con otros docentes y comparte" },
+  { icon: "rewarded_ads", title: "Reconocimientos", description: "Premios por participación y logros" },
+  { icon: "emoji_objects", title: "Ideas", description: "Inspírate con buenas prácticas educativas" },
+];
 
-const ArcoRecrea = () => {
+
+const ArcoRecrea = ({ children, displayHalf = false, isOnClickDisabled = false }) => {
+  const [selectedIndex, setSelectedIndex] = useState(null);
   return (
-    <div className="arco-container">
-      <svg viewBox="0 0 700 350" className="arco-svg">
-        <defs>
-          <linearGradient id="gradienteArco" x1="0%" y1="0%" x2="100%" y2="0%">
-            <stop offset="0%" stopColor="#fbbf24" />
-            <stop offset="25%" stopColor="#ec4899" />
-            <stop offset="50%" stopColor="#6366f1" />
-            <stop offset="75%" stopColor="#06b6d4" />
-            <stop offset="100%" stopColor="#84cc16" />
-          </linearGradient>
-        </defs>
-        <path
-          d="M100,350 A250,250 0 0,1 600,350"
-          stroke="url(#gradienteArco)"
-          strokeWidth="18"
-          fill="none"
-        />
-      </svg>
-
-      <div className="iconos-rotatorios">
-        {iconos.map((icon, i) => (
-          <div key={i} className={`icono icono-${i}`}>
-            <span className="material-symbols-outlined">{icon}</span>
+      <div className={ displayHalf ? 'cointainer' : ''}>
+        <div className="rotating-circle-container">
+          <div className="rotating-circle">
+            {iconos.map((item, index) => (
+              <div
+                key={index}
+                className={`icon ${selectedIndex === index ? 'active' : ''}`}
+                onClick={() => !isOnClickDisabled && setSelectedIndex(index)}
+              >
+                <span className="material-symbols-outlined">{item.icon}</span>
+              </div>
+            ))}
           </div>
-        ))}
-      </div>
 
-      <div className="texto-centro-arco">
-        <h1>Recrea <span>Academy</span></h1>
-        <p className="anio">2025</p>
+          <div className="center-circle">
+            <div className={`flip-card ${selectedIndex !== null ? 'flipped' : ''}`}>
+              <div className="flip-card-inner">
+                <div className="flip-card-front">
+                  {children}
+                </div>
+                <div className="flip-card-back">
+                  <Card className="card" sx={{ maxWidth: 300 }}>
+                    <div className="card-content">
+                      <h2>{selectedIndex !== null ? iconos[selectedIndex].title : ''}</h2>
+                      <p>{selectedIndex !== null ? iconos[selectedIndex].description : ''}</p>
+                    </div>
+                  </Card>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
       </div>
-    </div>
   );
 };
